Unit testing BudgetCategoryModel. Interim. Starting with Interface-Based (Strategy Pattern)
Moved the Db call, now I need to be able to specify change which Db call to be used. ie inject method.

Added a method to the DataTableModelEnum which gets data fron a file, and sets the internal data table. Move that to an implementation of IGetTableData

Later if ever
Then play with passing method pointers, Interfaces with the builder style (chat), and putting db calls in an enum too.