Unit testing BudgetCategoryModel. Interim. Starting with Interface-Based (Strategy Pattern) Tested GetDataFromFile. Now move it and re test. Fix the failed test
Done and Tested
Moved the Db call, now I need to be able to specify which Db call to be used. ie inject method.

Done and tested
Added a method to the DataTableModelEnum which gets data fron a file, and sets the internal data table. Still TBD...Move that to an implementation of IGetTableData

Later if ever
Then play with passing method pointers, Interfaces with the builder style (chat), and putting db calls in an enum too.